# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ord",
    "description": "◉ Ordinal wallet and block explorer",
    "license": {
      "name": "CC0-1.0"
    },
    "version": "0.14.1"
  },
  "paths": {
...
    "/api/v1/ord/block/{blockhash}/inscriptions": {
      "get": {
        "tags": [
          "ord"
        ],
        "summary": "Retrieve the inscription actions from the given block.",
        "description": "Retrieve the inscription actions from the given block.",
        "operationId": "ord_block_inscriptions",
        "parameters": [
          {
            "name": "blockhash",
            "in": "path",
            "description": "block hash",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Obtain inscription actions by blockhash",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiOrdBlockInscriptions"
                }
              }
            }
          },
          "400": {
            "description": "Bad query.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 2,
                  "msg": "bad request"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 3,
                  "msg": "not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1,
                  "msg": "internal error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ord/outpoint/{outpoint}/info": {
      "get": {
        "tags": [
          "ord"
        ],
        "summary": "Retrieve the outpoint infomation with the specified outpoint.",
        "description": "Retrieve the outpoint infomation with the specified outpoint.",
        "operationId": "ord_outpoint",
        "parameters": [
          {
            "name": "outpoint",
            "in": "path",
            "description": "Outpoint",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Obtain outpoint infomation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiOrdOutPointData"
                }
              }
            }
          },
          "400": {
            "description": "Bad query.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 2,
                  "msg": "bad request"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 3,
                  "msg": "not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1,
                  "msg": "internal error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ord/tx/{txid}/inscriptions": {
      "get": {
        "tags": [
          "ord"
        ],
        "summary": "Retrieve the inscription actions from the given transaction.",
        "description": "Retrieve the inscription actions from the given transaction.",
        "operationId": "ord_txid_inscriptions",
        "parameters": [
          {
            "name": "txid",
            "in": "path",
            "description": "transaction ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Obtain inscription actions by txid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiOrdTxInscriptions"
                }
              }
            }
          },
          "400": {
            "description": "Bad query.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 2,
                  "msg": "bad request"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 3,
                  "msg": "not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1,
                  "msg": "internal error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "Internal"
            ],
            "properties": {
              "Internal": {
                "type": "string",
                "description": "Internal server error."
              }
            },
            "example": {
              "code": 1,
              "msg": "internal error"
            }
          },
          {
            "type": "object",
            "required": [
              "BadRequest"
            ],
            "properties": {
              "BadRequest": {
                "type": "string",
                "description": "Bad request."
              }
            },
            "example": {
              "code": 1,
              "msg": "bad request"
            }
          },
          {
            "type": "object",
            "required": [
              "NotFound"
            ],
            "properties": {
              "NotFound": {
                "type": "string",
                "description": "Resource not found."
              }
            },
            "example": {
              "code": 1,
              "msg": "not found"
            }
          }
        ]
      },
...
      "ApiOrdBlockInscriptions": {
        "type": "object",
        "required": [
          "code",
          "msg",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "$ref": "#/components/schemas/ord.ApiBlockInscriptions"
          },
          "msg": {
            "type": "string",
            "description": "ok",
            "example": "ok"
          }
        }
      },
...
      "ApiOrdOutPointData": {
        "type": "object",
        "required": [
          "code",
          "msg",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "$ref": "#/components/schemas/ord.ApiOutpointInscriptions"
          },
          "msg": {
            "type": "string",
            "description": "ok",
            "example": "ok"
          }
        }
      },
...
      "ApiOrdTxInscriptions": {
        "type": "object",
        "required": [
          "code",
          "msg",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "$ref": "#/components/schemas/ord.ApiTxInscriptions"
          },
          "msg": {
            "type": "string",
            "description": "ok",
            "example": "ok"
          }
        }
      },
...
      "ScriptPubkey": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "address"
            ],
            "properties": {
              "address": {
                "type": "string",
                "description": "Address."
              }
            }
          },
          {
            "type": "object",
            "required": [
              "nonStandard"
            ],
            "properties": {
              "nonStandard": {
                "type": "string",
                "description": "Non-standard script hash."
              }
            }
          }
        ]
      },
...
      "ord.ApiBlockInscriptions": {
        "type": "object",
        "required": [
          "block"
        ],
        "properties": {
          "block": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ord.ApiTxInscriptions"
            }
          }
        }
      },
...
      "ord.ApiInscriptionAction": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "new"
            ],
            "properties": {
              "new": {
                "type": "object",
                "description": "New inscription",
                "required": [
                  "cursed",
                  "unbound"
                ],
                "properties": {
                  "cursed": {
                    "type": "boolean"
                  },
                  "unbound": {
                    "type": "boolean"
                  }
                }
              }
            }
          },
          {
            "type": "string",
            "description": "Transfer inscription",
            "enum": [
              "transfer"
            ]
          }
        ]
      },
      "ord.ApiInscriptionDigest": {
        "type": "object",
        "required": [
          "id",
          "number",
          "location"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The inscription id."
          },
          "location": {
            "type": "string",
            "description": "The inscription location."
          },
          "number": {
            "type": "integer",
            "format": "int32",
            "description": "The inscription number."
          }
        }
      },
...
      "ord.ApiOutpointInscriptions": {
        "type": "object",
        "required": [
          "txid",
          "scriptPubKey",
          "owner",
          "value",
          "inscriptionDigest"
        ],
        "properties": {
          "inscriptionDigest": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ord.ApiInscriptionDigest"
            },
            "description": "The inscriptions on the transaction output."
          },
          "owner": {
            "$ref": "#/components/schemas/ScriptPubkey"
          },
          "scriptPubKey": {
            "type": "string",
            "description": "The script pubkey."
          },
          "txid": {
            "type": "string",
            "description": "The transaction id."
          },
          "value": {
            "type": "integer",
            "format": "uint64",
            "description": "The value of the transaction output.",
            "minimum": 0
          }
        }
      },
      "ord.ApiTxInscription": {
        "type": "object",
        "required": [
          "action",
          "inscriptionId",
          "oldSatpoint",
          "from"
        ],
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ord.ApiInscriptionAction"
          },
          "from": {
            "$ref": "#/components/schemas/ScriptPubkey"
          },
          "inscriptionId": {
            "type": "string",
            "description": "The inscription id."
          },
          "inscriptionNumber": {
            "type": "integer",
            "format": "int32",
            "description": "The inscription number.",
            "nullable": true
          },
          "newSatpoint": {
            "type": "string",
            "description": "The inscription satpoint of the transaction output.",
            "nullable": true
          },
          "oldSatpoint": {
            "type": "string",
            "description": "The inscription satpoint of the transaction input."
          },
          "to": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ScriptPubkey"
              }
            ],
            "nullable": true
          }
        }
      },
      "ord.ApiTxInscriptions": {
        "type": "object",
        "required": [
          "inscriptions",
          "txid"
        ],
        "properties": {
          "inscriptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ord.ApiTxInscription"
            }
          },
          "txid": {
            "type": "string"
          }
        }
      }
    }
  }
}